Added check for hash before processing tile's bitstream
Now checks for an existing hash and uses the provided values before
again computing them. This is to allow short circuiting and avoiding
processing the bitstream in the future. At the moment the current
implementation causes a misalignment of the program and the bitstream
causing incorrect data to be read from the bitstream in future
iterations.

Signed-off-by: Sean Allesina-McGrory <allesis@wwu.edu>